backend/requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
pydantic-ai